+++ __tests__/unhappy.postActivityFatalAfterConnect.js
@@ -1,39 +1,15 @@
 import 'dotenv/config';
 
-import { createProxyServer } from 'http-proxy';
-import { createServer } from 'http';
-import { promisify } from 'util';
 import getPort from 'get-port';
 import onErrorResumeNext from 'on-error-resume-next';
 
 import { DirectLine } from '../src/directLine';
 import { timeouts } from './constants.json';
 import * as createDirectLineOptions from './setup/createDirectLineOptions';
+import createDirectLineForwarder from './setup/createDirectLineForwarder';
 import postActivity from './setup/postActivity';
 import waitForBotEcho from './setup/waitForBotEcho';
 import waitForConnected from './setup/waitForConnected';
-
-async function setupDirectLineForwarder(port, handler, target = 'https://directline.botframework.com/') {
-  // We need a reverse proxy (a.k.a. forwarder) to control the network traffic.
-  // This is because we need to modify the HTTP header by changing its host header (directline.botframework.com do not like "Host: localhost").
-
-  const proxy = createProxyServer({
-    changeOrigin: true,
-    rejectUnauthorized: false,
-    target
-  });
-
-  const proxyServer = createServer((req, res) => {
-    handler(req, res, () => proxy.web(req, res));
-  });
-
-  await promisify(proxyServer.listen.bind(proxyServer))(port);
-
-  return {
-    domain: `http://localhost:${ port }/v3/directline`,
-    unsubscribe: promisify(proxyServer.close.bind(proxyServer))
-  };
-}
 
 describe('Unhappy path', () => {
   let unsubscribes;
@@ -106,7 +82,7 @@ describe('Unhappy path', () => {
 
       let alwaysReturn404;
 
-      const { unsubscribe } = await setupDirectLineForwarder(proxyPort, (req, res, next) => {
+      const { unsubscribe } = await createDirectLineForwarder(proxyPort, (req, res, next) => {
         if (
           req.method !== 'OPTIONS'
           && alwaysReturn404
@@ -129,7 +105,7 @@ describe('Unhappy path', () => {
 
       alwaysReturn404 = true;
 
-      await expect(postActivity(directLine, { text: 'Hello, World!', type: 'message' })).rejects.toThrow();
+      await expect(postActivity(directLine, { text: 'Should not be sent', type: 'message' })).rejects.toThrow();
 
       // After post failed, it should stop polling and end all connections
